// src/models/management/core/operation/operationCostingDetail.interface.ts
type BaseDetailType = {
  id: number;
  costingId: number;
  supplierId: number;
  operationId: number;
  sellableId: number;
  sellableCode: string;
  paymentQuantity: number;
  amount: number;
  totalAmount: number;
  isLocked: boolean;
  lockedUpdate: string | null;
  isFinalized: boolean;
  finalizedUpdate: string | null;
  status: "NEW" | "DONE";
};
export type AirDetailType = {
  type: EStockType.AIRTICKET | EStockType.INSURANCE | EStockType.OTHER;
  details: {
    classOfService: string;
    tripType: "ONEWAY" | "ROUNDTRIP" | "MULTICITY";
    adult: number;
    child: number;
    infant: number;
    fullItinerary: string;
    departureDate: string;
    arrivalDate: string;
    specialRequest: string;
    remark: string;
  } | null;
};
// ...
export type ICostingDetailItem<T = void> = T & BaseDetailType;